pyproject.toml
[project]
name = "kensuiou-local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dlib>=20.0.0",
    "face-recognition>=1.3.0",
    "face-recognition-models",
    "gitpython>=3.1.44",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "psycopg2>=2.9.10",
    "pygame>=2.6.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "thop>=0.1.1.post2209072238",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.156",
    "websockets>=15.0.1",
]

[tool.uv.sources]
face-recognition-models = { git = "https://github.com/ageitgey/face_recognition_models" }
